refactor: refactored calendar.tsx and broke it down into smaller files for cleaner code
